Beginnings of data structure rewrite
I rewrote most of this project while I couldn't sleep and so a lot of it suffers from really poorly thought out design decisions. I started working on some QOL improvements for myself on running this weekly. Decided to rewrite my data structures in this

See notes in the README for more info