.Net 6 Conversion + Re-Arch
When setting up to run the first poll for the 2023 season I was frustrated by a couple of errors that popped up and unnecessary difficulty in fixing them due to how intertwined the code layers were. Static classes all over doing things they shouldn't be remotely responsible for. Yuck.

In the process of reorganizing everything I decided to also switch the project to .Net 6.

Ratings are at parity with the old version of the code but predictions aren't because I found a bug in them while fixing things up