2025 Cleanup Part 1
As with every year, in preparation for the upcoming season I decided to do some  ̶s̶p̶r̶i̶n̶g̶ summer cleaning.

I rearranged a lot of code (which makes most of the count of files changed) and removed almost everything related to betting (for now) as well as the temp UI code I had started on but decided I didn't like as it was.

I started a shift to make the code DB-driven. This exists right now as a SQL project with some initial tables and way more work to do. I started on a console app to extract the Excel file data into my tables. This whole thing should help automate running this throughout the season and save me time in the long run.

Once I have the data loaded in the DB then I can start to pivot the rest of the code to using it instead of the physical files.