2025 Cleanup Part 2
Added a couple new tables and have the population parts of the data loader utility working (as far as I can tell). It doesn't handle FCS teams yet so I'll need to deal with that at some point. Also the weeks in my data vs the "actual" weeks of the season are a little wonky and I'll need to reconcile that probably by creating a Week lookup to align them together.

Anyway, next steps will be purging the DB and doing a bigger load to make sure it is populating as I need it to. And once everything gets migrated I can start to pivot the rating code away from Excel.