2025 Cleanup Part 3
More work towards a DB driven design. The data loader now fully handles loading in teams for each season, their statistics, and game scores. It also has handling to put all FCS teams into a catch-all FCS group. There is still more work to do but it is meaningful progress towards moving the rest of the code to be DB driven.

Next steps are to do start building pathways in the code to pull the data from the DB instead of where it currently does and replace it.