Run Scenarios
Added the ability for the code to run and generate scenarios. The current output format is a little crude so need to come up with something easier to parse. That's future (tomorrow) me's problem when I analyze the scenarios I generated and add the playoff probability info in the ReadMe